fix(anthropic): harden SSE lifecycle
Reserve every streamed content index and reject unmatched stops so malformed streams cannot repair themselves. Preserve Anthropic citation payloads and combine framed transform errors with lifecycle finalization context.\n\nRefs #1237
