+++ src/core/providers/base/sse/cohere.rs
@@ -85,6 +85,7 @@ impl SSETransformer for CohereTransformer {
                                 tool_calls: None,
                                 function_call: None,
                                 audio: None,
+                                annotations: None,
                             },
                             finish_reason: None,
                             logprobs: None,
@@ -138,6 +139,7 @@ impl SSETransformer for CohereTransformer {
                                 tool_calls: None,
                                 function_call: None,
                                 audio: None,
+                                annotations: None,
                             },
                             finish_reason: Some(Self::parse_cohere_finish_reason(finish_reason)),
                             logprobs: None,
@@ -168,6 +170,7 @@ impl SSETransformer for CohereTransformer {
                                 tool_calls: None,
                                 function_call: None,
                                 audio: None,
+                                annotations: None,
                             },
                             finish_reason: None,
                             logprobs: None,
@@ -195,6 +198,7 @@ impl SSETransformer for CohereTransformer {
                                 tool_calls: None,
                                 function_call: None,
                                 audio: None,
+                                annotations: None,
                             },
                             finish_reason: Some(Self::parse_cohere_finish_reason(finish_reason)),
                             logprobs: None,
