+++ src/core/providers/replicate/provider.rs
@@ -474,6 +474,7 @@ impl LLMProvider for ReplicateProvider {
                         tool_calls: None,
                         function_call: None,
                         audio: None,
+                        annotations: None,
                     },
                     logprobs: None,
                     finish_reason: Some(crate::core::types::responses::FinishReason::Stop),
