+++ src/core/types/responses/chat.rs
@@ -402,6 +402,7 @@ mod tests {
                     tool_calls: None,
                     function_call: None,
                     audio: None,
+                    annotations: None,
                 },
                 finish_reason: None,
                 logprobs: None,
@@ -426,6 +427,7 @@ mod tests {
                 tool_calls: None,
                 function_call: None,
                 audio: None,
+                annotations: None,
             },
             finish_reason: Some(FinishReason::Stop),
             logprobs: None,
